Add telescope eod j2000 option (#48)
* Increase precision of telescope conversion constants.
* Add option to allow telescope to use either astrometric, J2000, or
  apparent coordinates at current epoch.

This work began from a request for a new feature from a user in South
Africa. We worked together both on the requirements and testing after I
sent him these changed files. I also tested against my telescope
simulator and both of us agreed these changes work well.

Along the way we discovered one tiny source of error was in these two
values used for the conversion of dec in degrees and RA in hours to
radians, so I computed them to insane accuracy and pasted them in. Now
that particular limitation will never appear again :-)